fix bug

+++ auto_bangumi/core/renamer.py
@@ -66,8 +66,9 @@ class Renamer:
                         self.client.rename_torrent_file(hash, path_name, new_name)
                         rename_count += 1
                 except:
-                    self.client.delete_torrent(info.hash)
-                    logger.warning(f"{name} rename fail")
+                    logger.warning(f"{info.name} rename fail")
+                    if settings.remove_bad_torrent:
+                        self.client.delete_torrent(info.hash)
             self.print_result(torrent_count, rename_count)
